Add users/groups; Env changes; Undo Rayon; Docs
A username and group name can optionally be included in a service config
file so that that service is run as a user other than root. These can be
defined on the service level (applies to every method), or
defined/overridden on the method level.

I also wrote up some documentation here, hopefully can help people
approaching the codebase for the first time.