Support Namespaces; Use syscalls to impl Command
While attempting to implement namespaces for services, I found that
std::process::Command was inadequate for the task, so this commit
includes a reimplementation of it with redox syscalls. It's not as good,
but it does what I need, and hopefully will be improved upon as I
continue to work on this.

Namespaces work now!