Add Process struct; Replace logger + colored logging
Using fern instead of simple_logger. Idk if it's a good idea, but I like
fern and it works well. I'm using standard colored too for output, I've
heard stories about it being slow, so we'll see if it's a problem. Might
change libs at some point, but it works well enough for now.